backend/requirements.txt
flask
flask-cors
werkzeug
flask-bcrypt
flask-jwt-extended
flask-cors
psycopg2-binary